% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anwObstacles.R
\docType{data}
\name{anwObstacles}
\alias{anwObstacles}
\title{American Ninja Warrior obstacles.}
\format{A data frame with 889 rows and 5 variables:
\describe{
  \item{Season}{Season of the show for given observation}
  \item{Location}{Global location of the obstacle for given obsevation}
  \item{Round/Stage}{What phase of the competition, within a given season, this obstacle was seen}
  \item{ObstacleName}{Name of the given obstacle}
  \item{ObstacleOrder}{Ordered placement of obstacle on the given course}
}}
\source{
\url{https://data.world/ninja/anw-obstacle-history}
}
\usage{
anwObstacles
}
\description{
A dataset of every obstacle used in American Ninja Warrior from seasons 1-10.
}
\keyword{datasets}
